requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
openai==1.54.0
firebase-admin==6.4.0
gunicorn==21.2.0
